不要なもの削除
