--- demo/src/main/java/com/example/demo/service/impl/ProductDetailServiceImpl.java
@@ -1,27 +1,0 @@
-package com.example.demo.service.impl;
-
-import java.util.Optional;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.example.demo.entity.ProductDetail;
-import com.example.demo.repository.ProductDetailRepository;
-import com.example.demo.service.ProductDetailService;
-
-@Service
-public class ProductDetailServiceImpl implements ProductDetailService {
-
-    @Autowired
-    private ProductDetailRepository productDetailRepository;
-
-    @Override
-    public Optional<ProductDetail> getProductDetailById(Long id) {
-        return productDetailRepository.findById(id);
-    }
-
-    @Override
-    public void saveProductDetail(ProductDetail productDetail) {
-        productDetailRepository.save(productDetail);
-    }
-}
